refactor(core): extract collector key indexing into KeyIndex
