+++ packages/core/src/data/collector.js
@@ -8,75 +8,7 @@ import { asArray } from "../utils/arrayUtils.js";
 import { radixSortIntoLookupArray } from "../utils/radixSort.js";
 import { UNIQUE_ID_KEY } from "./transforms/identifier.js";
 import createDomain from "../utils/domainArray.js";
-
-const MULTI_KEY_SEPARATOR = "|";
-const MULTI_KEY_ESCAPE = "\\";
-
-/**
- * @param {unknown} value
- * @param {string[]} keyFields
- * @param {number} index
- * @returns {import("../spec/channel.js").Scalar}
- */
-function validateKeyComponent(value, keyFields, index) {
-    const fieldName = keyFields[index];
-    if (value === undefined) {
-        throw new Error(
-            `Key field "${fieldName}" is undefined. Ensure all key fields are present in the data.`
-        );
-    }
-
-    if (value === null) {
-        throw new Error(
-            `Key field "${fieldName}" is null. Ensure all key fields are present in the data.`
-        );
-    }
-
-    if (
-        typeof value !== "string" &&
-        typeof value !== "number" &&
-        typeof value !== "boolean"
-    ) {
-        throw new Error(
-            `Key field "${fieldName}" must be a scalar value (string, number, or boolean).`
-        );
-    }
-
-    return value;
-}
-
-/**
- * @param {string} value
- * @returns {string}
- */
-function escapeKeyString(value) {
-    return value
-        .replaceAll(MULTI_KEY_ESCAPE, MULTI_KEY_ESCAPE + MULTI_KEY_ESCAPE)
-        .replaceAll(
-            MULTI_KEY_SEPARATOR,
-            MULTI_KEY_ESCAPE + MULTI_KEY_SEPARATOR
-        );
-}
-
-/**
- * @param {string[]} keyFields
- * @param {unknown[]} keyTuple
- * @returns {string}
- */
-function makeCompositeKey(keyFields, keyTuple) {
-    return keyTuple
-        .map((value, i) => {
-            const scalar = validateKeyComponent(value, keyFields, i);
-            if (typeof scalar === "string") {
-                return "s:" + escapeKeyString(scalar);
-            } else if (typeof scalar === "number") {
-                return "n:" + String(scalar);
-            } else {
-                return scalar ? "b:1" : "b:0";
-            }
-        })
-        .join(MULTI_KEY_SEPARATOR);
-}
+import KeyIndex from "./keyIndex.js";
 
 /**
  * Collects (materializes) the data that flows through this node.
@@ -103,14 +35,8 @@ export default class Collector extends FlowNode {
      */
     #uniqueIdIndex = [];
 
-    /** @type {Map<import("../spec/channel.js").Scalar | string, Datum> | null} */
-    #keyIndex = null;
-
-    /** @type {string[] | null} */
-    #keyIndexFields = null;
-
-    /** @type {boolean} */
-    #keyIndexUsesTuple = false;
+    /** @type {KeyIndex} */
+    #keyIndex = new KeyIndex();
     /**
      * Start and end indices of all facets if they are concatenated into a single array.
      * Used together with the uniqueIdIndex for looking up data items by their unique id.
@@ -157,7 +83,7 @@ export default class Collector extends FlowNode {
     #init() {
         this.#buffer = [];
         this.#uniqueIdIndex = [];
-        this.#invalidateKeyIndex();
+        this.#keyIndex.invalidate();
 
         this.facetBatches.clear();
         this.facetBatches.set(undefined, this.#buffer);
@@ -179,7 +105,7 @@ export default class Collector extends FlowNode {
      * @param {import("../types/flowBatch.js").FlowBatch} flowBatch
      */
     beginBatch(flowBatch) {
-        this.#invalidateKeyIndex();
+        this.#keyIndex.invalidate();
 
         if (isFacetBatch(flowBatch)) {
             this.#buffer = [];
@@ -411,83 +337,6 @@ export default class Collector extends FlowNode {
     }
 
     /**
-     * @param {string[]} keyFields
-     */
-    #buildKeyIndex(keyFields) {
-        /** @type {Array<(datum: Datum) => import("../spec/channel.js").Scalar>} */
-        const accessors = keyFields.map((fieldName) => field(fieldName));
-
-        /** @type {Map<import("../spec/channel.js").Scalar | string, Datum>} */
-        const index = new Map();
-
-        const useTuple = keyFields.length !== 1;
-
-        for (const data of this.facetBatches.values()) {
-            for (let i = 0, n = data.length; i < n; i++) {
-                const datum = data[i];
-                const keyTuple = accessors.map((accessor) => accessor(datum));
-                const key = useTuple
-                    ? makeCompositeKey(keyFields, keyTuple)
-                    : validateKeyComponent(keyTuple[0], keyFields, 0);
-
-                if (index.has(key)) {
-                    const duplicateValue = useTuple ? keyTuple : key;
-                    throw new Error(
-                        `Duplicate key detected for fields [${keyFields.join(
-                            ", "
-                        )}]: ${JSON.stringify(duplicateValue)}`
-                    );
-                }
-
-                index.set(key, datum);
-            }
-        }
-
-        this.#keyIndex = index;
-        this.#keyIndexFields = [...keyFields];
-        this.#keyIndexUsesTuple = useTuple;
-    }
-
-    /**
-     * @param {string[]} keyFields
-     * @returns {Map<import("../spec/channel.js").Scalar | string, Datum>}
-     */
-    #getKeyIndex(keyFields) {
-        if (!this.#keyIndex || !this.#matchesKeyFields(keyFields)) {
-            this.#buildKeyIndex(keyFields);
-        }
-
-        return this.#keyIndex;
-    }
-
-    /**
-     * @param {string[]} keyFields
-     */
-    #matchesKeyFields(keyFields) {
-        if (!this.#keyIndexFields) {
-            return false;
-        }
-
-        if (this.#keyIndexFields.length !== keyFields.length) {
-            return false;
-        }
-
-        for (let i = 0; i < keyFields.length; i++) {
-            if (this.#keyIndexFields[i] !== keyFields[i]) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    #invalidateKeyIndex() {
-        this.#keyIndex = null;
-        this.#keyIndexFields = null;
-        this.#keyIndexUsesTuple = false;
-    }
-
-    /**
      * Use an index to find a datum by its unique id.
      *
      * @param {number} uniqueId
@@ -528,24 +377,11 @@ export default class Collector extends FlowNode {
      */
     findDatumByKey(keyFields, keyTuple) {
         this.#checkStatus();
-
-        if (!keyFields || keyFields.length === 0) {
-            return;
-        }
-
-        if (keyFields.length !== keyTuple.length) {
-            throw new Error(
-                `Key tuple length ${keyTuple.length} does not match fields [${keyFields.join(
-                    ", "
-                )}]`
-            );
-        }
-
-        const index = this.#getKeyIndex(keyFields);
-        const key = this.#keyIndexUsesTuple
-            ? makeCompositeKey(keyFields, keyTuple)
-            : validateKeyComponent(keyTuple[0], keyFields, 0);
-        return index.get(key);
+        return this.#keyIndex.findDatum(
+            keyFields,
+            keyTuple,
+            this.facetBatches.values()
+        );
     }
 }
